// liber-web/src/app/modules/openapi/model/session.ts
/**
 * LIBER-API
 *
 * Contact: gianluca.fattarsi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserInfo } from './user-info';


/**
 * Session information
 */
export interface Session { 
    /**
     * Access token
     */
    accessToken: string;
    /**
     * User environment
     */
    environment?: string;
    /**
     * Refresh token
     */
    refreshToken: string;
    /**
     * User information
     */
    userInfo: UserInfo;
}