// liber-web/src/app/modules/openapi/model/user-login.ts
/**
 * LIBER-API
 *
 * Contact: gianluca.fattarsi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * User login information
 */
export interface UserLogin { 
    /**
     * User password
     */
    password: string;
    /**
     * User name
     */
    username: string;
}